Refactor the GPT code examples into files in the examples directory to improve code clarity, and other minor fixes
