+++ examples/connection_lint/pylama.txt
@@ -1,0 +1,19 @@
+extract_connection_info.py:4:1 E302 expected 2 blank lines, found 1 [pycodestyle]
+extract_connection_info.py:5:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:7:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:10:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:11:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:13:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:14:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:15:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:16:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:17:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:19:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:20:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:21:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:22:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:23:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:25:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:26:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:27:3 E111 indentation is not a multiple of 4 [pycodestyle]
+extract_connection_info.py:34:3 E111 indentation is not a multiple of 4 [pycodestyle]
