+++ examples/fibonacci/fibonacci.mrsh
@@ -1,0 +1,10 @@
+# func fibonacci(integer): integer in the set of fibonacci numbers
+
+This function calculates the nth fibonacci number, where n is provided to it and starts with 1.
+
+fibonacci(n) = fibonacci(n - 1) + fibonacci(n - 2)
+
+* fibonacci(1) = 1
+* fibonacci(2) = 1
+* fibonacci(3) = 2
+* fibonacci(0) throws an error
