+++ examples/test/extract_connection_info.mrsh
